* enc/{emacs_mule,euc_jp}.c (code_to_mbc): suppress warnings. * enc/iso_8859_{1,2}.c (apply_all_case_fold): ditto.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@30468 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ enc/emacs_mule.c
@@ -264,7 +264,7 @@ code_to_mbc(OnigCodePoint code, UChar *buf, OnigEncoding enc)
 
   if (enclen(enc, buf, p) != (p - buf))
     return ONIGERR_INVALID_CODE_POINT_VALUE;
-  return p - buf;
+  return (int)(p - buf);
 }
 
 static int
